pyproject.toml
[project]
name = "personal-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi[standard]>=0.135.1",
    "huggingface-hub>=1.7.1",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.1",
    "openai>=2.28.0",
    "pinecone-client>=6.0.0",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "tiktoken>=0.12.0",
    "unstructured>=0.21.5",
]
